+++ persistence/src/main/java/com/epam/esm/persistence/mapper/OrderMapper.java
@@ -4,11 +4,10 @@ import com.epam.esm.model.entity.Certificate;
 import com.epam.esm.model.entity.Order;
 import com.epam.esm.model.entity.User;
 import com.epam.esm.persistence.constants.OrderColumns;
-import com.epam.esm.persistence.constants.UserColumns;
-import com.epam.esm.persistence.dao.impl.CertificateDAOImpl;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.time.ZoneId;
@@ -22,7 +21,7 @@ public class OrderMapper implements RowMapper<Order> {
         User user = new UserMapper().mapRow(resultSet,i);
         user.setId(resultSet.getInt(OrderColumns.CERTIFICATE_ID.getValue()));
         Order order = new Order(certificate, user,
-        resultSet.getFloat(OrderColumns.AMOUNT.getValue()),
+        BigDecimal.valueOf(resultSet.getFloat(OrderColumns.AMOUNT.getValue())),
                 resultSet.getInt(OrderColumns.QUANTITY.getValue()),
         resultSet.getTimestamp(OrderColumns.PURCHASE_DATE.getValue())
                 .toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime());
